refactor: move loading to ui foulder
